manager state: a corrupt bound is refused by name, not indexed past
Three bounds checks were computed in the width they were guarding. A table
naming a block whose offset and length are each a legal length but overrun
added together passed the check and then indexed out of the array; a
compressed literal declaring four bytes of length above the signed range
became a negative one, which both guards let through before the slice threw.
Each left the reader as a raw platform exception, past the composition point
that catches its own refusals and nothing else.

The additions are done in a width that holds them, and the literal's
declared run is refused before it is narrowed. Both sites carry a check
built from bytes rather than from a fixture.

+++ src/ripperdoc-core/ManagerState/TableFile.cs
@@ -106,7 +106,8 @@ internal static class TableFile
 
     private static byte[] Block(byte[] data, int offset, int length, string what)
     {
-        if (offset < 0 || length < 0 || offset + length + BlockTrailerSize > data.Length)
+        if (offset < 0 || length < 0
+            || (long)offset + length + BlockTrailerSize > data.Length)
         {
             throw new StateReadException(
                 $"'{what}' names a block of {length} bytes at byte {offset}, which runs past the "
@@ -176,7 +177,7 @@ internal static class TableFile
                     + "is corrupt.");
             }
 
-            if (at + fresh + valueLength > limit)
+            if ((long)at + fresh + valueLength > limit)
             {
                 throw new StateReadException(
                     $"'{what}' holds an entry whose key and value run past the end of its block. "
